+++ jdaviz/core/loaders/resolvers/url/url.vue
@@ -3,6 +3,7 @@
     title="Download from URL"
     :popout_button="popout_button"
     :spinner="spinner"
+    :parsed_input_is_resolvable="parsed_input_is_resolvable"
     :parsed_input_is_query="parsed_input_is_query"
     :treat_table_as_query.sync="treat_table_as_query"
     :observation_table="observation_table"
@@ -29,6 +30,7 @@
         style="padding: 0px 8px"
         :label="api_hints_enabled ? 'ldr.url =' : ''"
         :class="api_hints_enabled ? 'api-hint' : null"
+        :error-messages="parsed_input_is_resolvable ? [parsed_input_is_resolvable] : []"
       ></v-text-field>
     </v-row>