+++ jdaviz/core/loaders/importers/test_importer.py
@@ -179,14 +179,14 @@ def test_reject_2d_spectrum_as_image(deconfigged_helper, spectrum2d, mos_spectru
     Test that 2D spectra being read in as images are rejected.
     """
     # Attempt to load should raise a helpful error
-    with pytest.raises(ValueError, match="'object > Image': 'not valid'"):
+    with pytest.raises(ValueError, match="No valid loaders found for input."):
         deconfigged_helper.load(spectrum2d, format='Image')
 
     # Verify no data was loaded
     assert len(deconfigged_helper.app.data_collection) == 0
 
     # Try again with 2D Spectrum as HDU with spectral wcs
-    with pytest.raises(ValueError, match="'object > Image': 'not valid'"):
+    with pytest.raises(ValueError, match="No valid loaders found for input."):
         deconfigged_helper.load(mos_spectrum2d_as_hdulist[1], format='Image')
 
     # Verify no data was loaded
@@ -195,6 +195,30 @@ def test_reject_2d_spectrum_as_image(deconfigged_helper, spectrum2d, mos_spectru
     # Verify we're able to load as 2d spectrum in follow-up
     deconfigged_helper.load(spectrum2d, format='2D Spectrum')
     assert len(deconfigged_helper.app.data_collection) > 0
+
+
+def test_image_importer_expose_align_by_options(deconfigged_helper, image_hdu_wcs):
+    """
+    Verify that the logic for expose_align_by_options, the traitlet in the image
+    importer that determines whether or not to expose option to align by pixels
+    or WCS, is correct. It should be True (to show options) if there is an
+    Orientation plugin or if the selection for the new viewer creation is 'Image',
+    and False otherwise.
+    """
+
+    ldr = deconfigged_helper.loaders['object']
+    ldr.object = image_hdu_wcs
+    ldr.format = 'Image'
+
+    importer = ldr.importer._obj
+
+    # selecting Image viewer for create new should expose options
+    importer.viewer.create_new.selected = 'Image'
+    assert importer.expose_align_by_options is True
+
+    # clearing create new Image viewer selection should hide options again
+    importer.viewer.create_new.selected = ''
+    assert importer.expose_align_by_options is False
 
 
 class TestDataLabelOverwrite: