+++ CHANGES.rst
@@ -130,6 +130,8 @@ Specviz
 Specviz2d
 ^^^^^^^^^
 
+- Fixed bug where loading two 2D spectra failed to display in the spectrum-2d viewer. [#3983]
+
 4.5.1 (2026-03-06)
 ==================
 
@@ -186,6 +188,7 @@ Specviz
 
 Specviz2d
 ^^^^^^^^^
+
 - Fixed an issue with GWCS 1.0.3 by updating the ``try/except`` to an if check against ``pixel_n_dim``. [#4032]
 
 4.5 (2025-12-15)